Add story
Split Tooltip stories into a separate file to add structure to Storybook
