+++ projects/js-packages/charts/src/components/line-chart/stories/tooltip.stories.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import LineChart from '../line-chart';
+import { lineChartMetaArgs, lineChartStoryArgs } from './config';
+import type { Meta, StoryFn, StoryObj } from '@storybook/react';
+
+const meta: Meta< typeof LineChart > = {
+	...lineChartMetaArgs,
+	title: 'JS Packages/Charts/Types/Line Chart/Tooltips',
+} satisfies Meta< typeof LineChart >;
+
+export default meta;
+
+const Template: StoryFn< typeof LineChart > = args => <LineChart { ...args } />;
+
+const tooltipStoryArgs = {
+	...lineChartStoryArgs,
+	withTooltipCrosshairs: false,
+};
+
+export const Default: StoryObj< typeof LineChart > = Template.bind( {} );
+Default.args = {
+	...tooltipStoryArgs,
+};
+
+export const NoTooltips: StoryObj< typeof LineChart > = Template.bind( {} );
+NoTooltips.args = {
+	...tooltipStoryArgs,
+	withTooltips: false,
+};
+
+export const Crosshairs: StoryObj< typeof LineChart > = Template.bind( {} );
+Crosshairs.args = {
+	...tooltipStoryArgs,
+	withTooltipCrosshairs: true,
+};
+
+export const Custom: StoryObj< typeof LineChart > = Template.bind( {} );
+Custom.args = {
+	...tooltipStoryArgs,
+	renderTooltip: ( { tooltipData } ) => {
+		const nearestDatum = tooltipData?.nearestDatum?.datum;
+		if ( ! nearestDatum ) return null;
+
+		const tooltipPoints = Object.entries( tooltipData?.datumByKey || {} )
+			.map( ( [ key, { datum } ] ) => ( {
+				key,
+				value: datum.value as number,
+			} ) )
+			.sort( ( a, b ) => b.value - a.value );
+
+		return (
+			<div>
+				<h3>{ nearestDatum?.date?.toLocaleDateString() } 💯 </h3>
+
+				<table style={ { border: '1px solid black', borderCollapse: 'collapse' } }>
+					{ tooltipPoints.map( point => (
+						<tr style={ { border: '1px solid black' } } key={ point.key }>
+							<td style={ { border: '1px solid black' } }>{ point.key }</td>
+							<td>{ point.value }</td>
+						</tr>
+					) ) }
+				</table>
+			</div>
+		);
+	},
+};
